restructure, test environ_get_first
