+++ README.md
@@ -54,6 +54,12 @@ See the top level `Makefile` and `.circleci/config.yml` for the development flow
 
 ## Changelog
 
+### 0.2.1
+
+#### Bugfixes
+
+- remove `pytest-xdist` from runtime dependencies
+
 ### 0.2.0
 
 #### Features
