+++ pytest_jobserver/system.py
@@ -7,17 +7,17 @@ FileDescriptor = NewType("FileDescriptor", int)
 FileDescriptorsRW = Tuple[FileDescriptor, FileDescriptor]
 
 
+def environ_get_first(keys: Iterable[str]) -> Optional[str]:
+    for key in keys:
+        value = os.environ.get(key)
+        if value is not None:
+            return value
+    return None
+
+
 def is_fifo(path: str) -> bool:
     return stat.S_ISFIFO(os.stat(path).st_mode)
 
 
 def is_rw_ok(path: str) -> bool:
     return os.access(path, os.R_OK | os.W_OK)
-
-
-def first_environ_get(keys: Iterable[str]) -> Optional[str]:
-    for key in keys:
-        value = os.environ.get(key)
-        if value is not None:
-            return value
-    return None
